Hide existing children when accordion initializes
Under normal DOM parsing, the ay-accordion gets initialized before its
child elements and they correctly trigger the MutationObserver and get
marked `hidden` as they initialize.

But there are cases (such as when appending a document fragment, or HTML
string, or when combined with various frameworks) that the ay-accordion
element initializes pre-populated with children and they do not trigger
the MutationObserver and do not get marked as hidden.